requirements.txt
Django==3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
sqlparse==0.3.0
django-appconf==1.0.3
python-memcached==1.59
PyYAML==5.4.1
django-cors-headers==3.7.0
Pillow==8.2.0
celery==5.0.5
redis==3.5.3